Fix IslandCustomDataQuery + Fix /is level scan + Ajout progress bar pour le /is level scan

+++ plugin/src/main/java/fr/euphyllia/skyllia/database/mariadb/MariaDBIslandCustomData.java
@@ -87,6 +87,7 @@ public class MariaDBIslandCustomData extends IslandCustomDataQuery {
 
     @Override
     public @Nullable <P, C> C get(@NotNull NamespacedKey namespace, @NotNull Island island, @NotNull String dataKey, @NotNull PersistentDataType<P, C> type) {
+        if (!ensureTableExists(namespace)) return null;
         String tableName = getTableName(namespace);
         String sql = String.format("""
                 SELECT data_value FROM %s
@@ -115,6 +116,7 @@ public class MariaDBIslandCustomData extends IslandCustomDataQuery {
 
     @Override
     public boolean has(@NotNull NamespacedKey namespace, @NotNull Island island, @NotNull String dataKey) {
+        if (!ensureTableExists(namespace)) return false;
         String tableName = getTableName(namespace);
         String sql = String.format("""
                 SELECT 1 FROM %s
@@ -144,6 +146,7 @@ public class MariaDBIslandCustomData extends IslandCustomDataQuery {
 
     @Override
     public @NotNull Set<String> getKeys(@NotNull NamespacedKey namespace, @NotNull Island island) {
+        if (!ensureTableExists(namespace)) return new HashSet<>();
         String tableName = getTableName(namespace);
         String sql = String.format("""
                 SELECT data_key FROM %s
@@ -166,6 +169,7 @@ public class MariaDBIslandCustomData extends IslandCustomDataQuery {
 
     @Override
     public boolean remove(@NotNull NamespacedKey namespace, @NotNull Island island, @NotNull String dataKey) {
+        if (!ensureTableExists(namespace)) return false;
         String tableName = getTableName(namespace);
         String sql = String.format("""
                 DELETE FROM %s
@@ -178,6 +182,7 @@ public class MariaDBIslandCustomData extends IslandCustomDataQuery {
 
     @Override
     public boolean clear(@NotNull NamespacedKey namespace, @NotNull Island island) {
+        if (!ensureTableExists(namespace)) return false;
         String tableName = getTableName(namespace);
         String sql = String.format("""
                 DELETE FROM %s
@@ -190,6 +195,7 @@ public class MariaDBIslandCustomData extends IslandCustomDataQuery {
 
     @Override
     public int size(@NotNull NamespacedKey namespace, @NotNull Island island) {
+        if (!ensureTableExists(namespace)) return 0;
         String tableName = getTableName(namespace);
         String sql = String.format("""
                 SELECT COUNT(*) FROM %s
